IGNITE-20335 Move IgniteRunner to internal package
