+++ packaging/build.gradle
@@ -67,7 +67,7 @@ def tokens = [
         CONF_DIR            : '${IGNITE_HOME}/etc',
         LIB_DIR             : '${IGNITE_HOME}/lib',
         APP_JAR             : "${project(':ignite-runner').name}-${project(':ignite-runner').version}.jar".toString(),
-        MAIN_CLASS          : 'org.apache.ignite.app.IgniteRunner',
+        MAIN_CLASS          : 'org.apache.ignite.internal.app.IgniteRunner',
         BOOTSTRAP_FILE_NAME : 'bootstrap-functions.sh',
         SETUP_JAVA_FILE_NAME: 'setup-java.sh'
 ]
